requirements.txt
astroid==3.3.11
blinker==1.9.0
branca==0.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
flask==3.1.1
folium==0.20.0
fonttools==4.58.5
git-filter-repo==2.47.0
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mypy==1.17.0
mypy-extensions==1.1.0
numpy==2.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
pip==24.0
platformdirs==4.3.8
pylint==3.3.7
pymysql==1.1.1
pyparsing==3.2.3
pyserial==3.5
pyside6==6.9.1
pyside6-addons==6.9.1
pyside6-essentials==6.9.1
pyside6-stubs==6.7.3.0
python-dateutil==2.9.0.post0
requests==2.32.4
shiboken6==6.9.1
six==1.17.0
tomlkit==0.13.3
typing-extensions==4.14.1
urllib3==2.5.0
uv==0.7.21
vtk==9.5.0
werkzeug==3.1.3
xyzservices==2025.4.0
